adjust to changed `HomekitAuthInfo` interface
HAP-Java changed the interface to have a `hasUser` method, which indicated whether a user was paired already or not.
This interface change remained undetected, because it has a default implementation.

The effect of not implementing the method is that HAP-Java will always be in pairing mode. A restart of the app will lead to 'no response' displayed in the home app.